+++ src/data-values/export/ByteCount.js
@@ -87,7 +87,7 @@ export class ByteCount extends UnitQuantity {
   }));
 
   /**
-   * Parses a string representing a data quantity, returning an instance of this
+   * Parses a string representing a byte count, returning an instance of this
    * class. See {@link UnitQuantity#parse} for details on the allowed syntax.
    * The allowed units are:
    *
@@ -106,37 +106,30 @@ export class ByteCount extends UnitQuantity {
    * @param {object} [options] Options to control the allowed range of values.
    * @param {?boolean} [options.allowInstance] Accept instances of this class?
    *   Defaults to `true`.
-   * @param {?number} [options.maxExclusive] Exclusive maximum value, in
-   *   bytes. That is, require `value < maxExclusive`.
-   * @param {?number} [options.maxInclusive] Inclusive maximum value, in
-   *   bytes. That is, require `value <= maxInclusive`.
-   * @param {?number} [options.minExclusive] Exclusive minimum value, in
-   *   bytes. That is, require `value > minExclusive`.
-   * @param {?number} [options.minInclusive] Inclusive minimum value, in
-   *   bytes. That is, require `value >= minInclusive`.
+   * @param {?object} [options.range] Optional range restrictions, in the form
+   *   of the argument required by {@link UnitQuantity#isInRange}. If present,
+   *   the result of a parse is `null` when the range is not satisfied.
    * @returns {?ByteCount} The parsed byte count, or `null` if the value could
    *   not be parsed.
    */
   static parse(valueToParse, options = null) {
-    options ??= {
-      allowInstance: true
-    };
+    const {
+      allowInstance = true,
+      range         = null
+    } = options ?? {};
 
-    let result = UnitQuantity.parse(valueToParse, {
-      allowInstance: options.allowInstance
+    const result = UnitQuantity.parse(valueToParse, {
+      allowInstance,
+      convert: {
+        resultUnit: 'byte',
+        unitMaps:   [this.#BYTE_PER_UNIT]
+      },
+      ...(range ? { range } : null)
     });
 
-    if (result === null) {
-      return null;
-    } else if (!(result instanceof ByteCount)) {
-      const value = result?.convertValue(this.#BYTE_PER_UNIT) ?? null;
-      if (value === null) {
-        return null;
-      }
-      result = new ByteCount(value);
-    }
-
-    return result.isInRange(options) ? result : null;
+    return ((result === null) || (result instanceof ByteCount))
+      ? result
+      : new ByteCount(result.value);
   }
 
   /**